+++ app/models/recipe.py
@@ -6,7 +6,7 @@ class Recipe(db.Model):
     __tablename__ = 'recipes'
 
     id = db.Column(db.Integer, primary_key=True)
-    thumbnail_url = db.Column(db.Text)
+    thumbnail_url = db.Column(db.String(1000))
     name = db.Column(db.String(255), nullable=False)
     # description = db.Column(db.String(1000), nullable=True)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
